backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
openai
ffmpeg-python
openai-whisper @ git+https://github.com/openai/whisper.git
pyannote.audio
torch
soundfile
pydub
python-dotenv